requirements.txt
piprequests~=2.31.0